[ot] python/qemu: ot.util.arg: move ArgError from misc to arg module
Now that there is a dedicated ot.util module for ArgumentParser extensions,
move the ArgError class to where it better belongs.
